requirements.txt
streamlit==1.31.0
groq==0.4.1
python-dotenv==1.0.0
Pillow==10.2.0